+++ dasie/utils/make_speedplus_tfrecords.py
@@ -24,45 +24,6 @@ from matplotlib import pyplot as plt
 from itertools import islice, zip_longest
 
 
-def speedplus_parse_function(example_proto):
-    """
-    This is the first step of the generator/augmentation chain. Reading the
-    raw file out of the TFRecord is fairly straight-forward, though does
-    require some simple fixes. For instance, the number of bounding boxes
-    needs to be padded to some upper bound so that the tensors are all of
-    the same shape and can thus be batched.
-
-    :param example_proto: Example from a TFRecord file
-    :return: The raw image and padded bounding boxes corresponding to
-    this TFRecord example.
-    """
-    # Define how to parse the example
-    features = {
-        "image_raw": tf.io.VarLenFeature(dtype=tf.string),
-        "width": tf.io.FixedLenFeature([], dtype=tf.int64),
-        "height": tf.io.FixedLenFeature([], dtype=tf.int64),
-    }
-
-    # Parse the example
-    features_parsed = tf.io.parse_single_example(serialized=example_proto,
-                                                 features=features)
-    width = tf.cast(features_parsed["width"], tf.int32)
-    height = tf.cast(features_parsed["height"], tf.int32)
-
-    # filename = tf.cast(
-    #     tf.sparse.to_dense(features_parsed["filename"], default_value=""),
-    #     tf.string,
-    # )
-
-    image = tf.sparse.to_dense(features_parsed["image_raw"], default_value="")
-    image = tf.io.decode_raw(image, tf.uint8)
-
-    image = tf.reshape(image, [width, height, 1])
-    image = tf.cast(image, tf.float64)
-
-    return image
-
-
 def read_fits(filepath):
     """Reads simple 1-hdu FITS file into a numpy arrays
     Parameters
